#2018 Close inpuststream instead of connection.
