+++ client/src/main/java/com/alibaba/nacos/client/config/impl/HttpSimpleClient.java
@@ -17,12 +17,12 @@ package com.alibaba.nacos.client.config.impl;
 
 import com.alibaba.nacos.api.common.Constants;
 import com.alibaba.nacos.api.exception.NacosException;
-import com.alibaba.nacos.client.config.utils.IOUtils;
+import com.alibaba.nacos.common.utils.IoUtils;
 import com.alibaba.nacos.client.config.utils.MD5;
 import com.alibaba.nacos.client.utils.ParamUtil;
 import com.alibaba.nacos.common.constant.HttpHeaderConsts;
-import com.alibaba.nacos.common.util.UuidUtils;
-import com.alibaba.nacos.common.util.VersionUtils;
+import com.alibaba.nacos.common.utils.UuidUtils;
+import com.alibaba.nacos.common.utils.VersionUtils;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -38,7 +38,6 @@ import java.util.Map;
  * Http tool
  *
  * @author Nacos
- *
  */
 public class HttpSimpleClient {
 
@@ -55,10 +54,10 @@ public class HttpSimpleClient {
         HttpURLConnection conn = null;
 
         try {
-            conn = (HttpURLConnection)new URL(url).openConnection();
+            conn = (HttpURLConnection) new URL(url).openConnection();
             conn.setRequestMethod("GET");
             conn.setConnectTimeout(ParamUtil.getConnectTimeout() > 100 ? ParamUtil.getConnectTimeout() : 100);
-            conn.setReadTimeout((int)readTimeoutMs);
+            conn.setReadTimeout((int) readTimeoutMs);
             List<String> newHeaders = getHeaders(url, headers, paramValues);
             setHeaders(conn, newHeaders, encoding);
 
@@ -68,14 +67,14 @@ public class HttpSimpleClient {
             String resp = null;
 
             if (HttpURLConnection.HTTP_OK == respCode) {
-                resp = IOUtils.toString(conn.getInputStream(), encoding);
+                resp = IoUtils.toString(conn.getInputStream(), encoding);
             } else {
-                resp = IOUtils.toString(conn.getErrorStream(), encoding);
+                resp = IoUtils.toString(conn.getErrorStream(), encoding);
             }
             return new HttpResult(respCode, conn.getHeaderFields(), resp);
         } finally {
-            if (conn != null) {
-                conn.disconnect();
+            if (conn != null && conn.getInputStream() != null) {
+                conn.getInputStream().close();
             }
         }
     }
@@ -111,10 +110,10 @@ public class HttpSimpleClient {
         }
         HttpURLConnection conn = null;
         try {
-            conn = (HttpURLConnection)new URL(url).openConnection();
+            conn = (HttpURLConnection) new URL(url).openConnection();
             conn.setRequestMethod("POST");
             conn.setConnectTimeout(ParamUtil.getConnectTimeout() > 3000 ? ParamUtil.getConnectTimeout() : 3000);
-            conn.setReadTimeout((int)readTimeoutMs);
+            conn.setReadTimeout((int) readTimeoutMs);
             conn.setDoOutput(true);
             conn.setDoInput(true);
             List<String> newHeaders = getHeaders(url, headers, paramValues);
@@ -126,14 +125,14 @@ public class HttpSimpleClient {
             String resp = null;
 
             if (HttpURLConnection.HTTP_OK == respCode) {
-                resp = IOUtils.toString(conn.getInputStream(), encoding);
+                resp = IoUtils.toString(conn.getInputStream(), encoding);
             } else {
-                resp = IOUtils.toString(conn.getErrorStream(), encoding);
+                resp = IoUtils.toString(conn.getErrorStream(), encoding);
             }
             return new HttpResult(respCode, conn.getHeaderFields(), resp);
         } finally {
-            if (null != conn) {
-                conn.disconnect();
+            if (conn != null && conn.getInputStream() != null) {
+                conn.getInputStream().close();
             }
         }
     }
@@ -167,10 +166,10 @@ public class HttpSimpleClient {
         HttpURLConnection conn = null;
 
         try {
-            conn = (HttpURLConnection)new URL(url).openConnection();
+            conn = (HttpURLConnection) new URL(url).openConnection();
             conn.setRequestMethod("DELETE");
             conn.setConnectTimeout(ParamUtil.getConnectTimeout() > 100 ? ParamUtil.getConnectTimeout() : 100);
-            conn.setReadTimeout((int)readTimeoutMs);
+            conn.setReadTimeout((int) readTimeoutMs);
             List<String> newHeaders = getHeaders(url, headers, paramValues);
             setHeaders(conn, newHeaders, encoding);
 
@@ -180,14 +179,14 @@ public class HttpSimpleClient {
             String resp = null;
 
             if (HttpURLConnection.HTTP_OK == respCode) {
-                resp = IOUtils.toString(conn.getInputStream(), encoding);
+                resp = IoUtils.toString(conn.getInputStream(), encoding);
             } else {
-                resp = IOUtils.toString(conn.getErrorStream(), encoding);
+                resp = IoUtils.toString(conn.getErrorStream(), encoding);
             }
             return new HttpResult(respCode, conn.getHeaderFields(), resp);
         } finally {
-            if (conn != null) {
-                conn.disconnect();
+            if (conn != null && conn.getInputStream() != null) {
+                conn.getInputStream().close();
             }
         }
     }
