+++ client/src/main/java/com/alibaba/nacos/client/logging/logback/LogbackNacosLogging.java
@@ -18,7 +18,7 @@ package com.alibaba.nacos.client.logging.logback;
 import ch.qos.logback.classic.LoggerContext;
 import ch.qos.logback.classic.util.ContextInitializer;
 import com.alibaba.nacos.client.logging.AbstractNacosLogging;
-import com.alibaba.nacos.common.util.ResourceUtils;
+import com.alibaba.nacos.common.utils.ResourceUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.impl.StaticLoggerBinder;
