+++ client/src/main/java/com/alibaba/nacos/client/naming/core/PushReceiver.java
@@ -16,7 +16,7 @@
 package com.alibaba.nacos.client.naming.core;
 
 import com.alibaba.fastjson.JSON;
-import com.alibaba.nacos.client.naming.utils.IoUtils;
+import com.alibaba.nacos.common.utils.IoUtils;
 import com.alibaba.nacos.client.utils.StringUtils;
 
 import java.net.DatagramPacket;
