+++ client/src/main/java/com/alibaba/nacos/client/naming/net/HttpClient.java
@@ -16,7 +16,7 @@
 package com.alibaba.nacos.client.naming.net;
 
 import com.alibaba.nacos.api.common.Constants;
-import com.alibaba.nacos.client.naming.utils.IoUtils;
+import com.alibaba.nacos.common.utils.IoUtils;
 import com.google.common.net.HttpHeaders;
 import org.apache.commons.lang3.StringUtils;
 
@@ -104,9 +104,7 @@ public class HttpClient {
 
             return new HttpResult(500, e.toString(), Collections.<String, String>emptyMap());
         } finally {
-            if (conn != null) {
-                conn.disconnect();
-            }
+            IoUtils.closeQuietly(conn);
         }
     }
