+++ common/src/main/java/com/alibaba/nacos/common/utils/Pair.java
@@ -13,16 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.nacos.common.util;
+package com.alibaba.nacos.common.utils;
 
-import java.util.UUID;
+import org.apache.commons.lang3.StringUtils;
 
 /**
- * @author nkorange
+ * @author nacos
  */
-public class UuidUtils {
+public class Pair {
+    private String key;
+    private String value;
 
-    public static String generateUuid() {
-        return UUID.randomUUID().toString();
+    public Pair(String key, String value) {
+        this.key = key;
+        this.value = value;
+    }
+
+    public Pair() {
+        this(StringUtils.EMPTY, StringUtils.EMPTY);
+    }
+
+    public String getKey() {
+        return key;
+    }
+
+    public void setKey(String key) {
+        this.key = key;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
     }
 }
