+++ common/src/main/java/com/alibaba/nacos/common/utils/ResourceUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.nacos.common.util;
+package com.alibaba.nacos.common.utils;
 
 import java.io.File;
 import java.io.FileNotFoundException;
