+++ common/src/main/java/com/alibaba/nacos/common/utils/UuidUtils.java
@@ -13,27 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.nacos.common.util;
+package com.alibaba.nacos.common.utils;
+
+import java.util.UUID;
 
 /**
  * @author nkorange
- * @since 0.8.0
  */
-public class HttpMethod {
+public class UuidUtils {
 
-    public static final String GET = "GET";
-
-    public static final String HEAD = "HEAD";
-
-    public static final String POST = "POST";
-
-    public static final String PUT = "PUT";
-
-    public static final String PATCH = "PATCH";
-
-    public static final String DELETE = "DELETE";
-
-    public static final String OPTIONS = "OPTIONS";
-
-    public static final String TRACE = "TRACE";
+    public static String generateUuid() {
+        return UUID.randomUUID().toString();
+    }
 }
